+++ fofix/core/VideoPlayer/__init__.py
@@ -1,1 +1,0 @@
-from fofix.lib._VideoPlayer import *